p2p/discover: Delete unused files Table and Discovery
