--- networks/p2p/discover/discover_storage.go
@@ -1,68 +1,0 @@
-// Modifications Copyright 2024 The Kaia Authors
-// Copyright 2019 The klaytn Authors
-// This file is part of the klaytn library.
-//
-// The klaytn library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The klaytn library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the klaytn library. If not, see <http://www.gnu.org/licenses/>.
-// Modified and improved for the Kaia development.
-
-package discover
-
-import (
-	"github.com/kaiachain/kaia/common"
-)
-
-type discoverStorage interface {
-	name() string
-	setTable(t *Table)
-	setTargetNodeType(tType NodeType)
-	init()
-
-	add(n *Node)
-	delete(n *Node)
-	len() (n int)
-	nodeAll() []*Node
-	readRandomNodes(buf []*Node) (n int)
-	closest(target common.Hash, nresults int) *nodesByDistance
-	stuff(nodes []*Node)
-	copyBondedNodes()
-
-	lookup(targetID NodeID, refreshIfEmpty bool, targetType NodeType) []*Node
-	getNodes(max int) []*Node
-	doRevalidate()
-	doRefresh()
-
-	isAuthorized(id NodeID) bool
-
-	// API
-	getBucketEntries() []*Node
-	getAuthorizedNodes() []*Node
-	putAuthorizedNode(node *Node)
-	deleteAuthorizedNode(id NodeID)
-}
-
-// nodeTypeName converts NodeType to string.
-func nodeTypeName(nt NodeType) string { // TODO-Kaia-Node Consolidate p2p.NodeType and common.ConnType
-	switch nt {
-	case NodeTypeCN:
-		return "CN"
-	case NodeTypePN:
-		return "PN"
-	case NodeTypeEN:
-		return "EN"
-	case NodeTypeBN:
-		return "BN"
-	default:
-		return "Unknown Node Type"
-	}
-}
